api/requirements.txt
fastapi==0.115.6
uvicorn==0.30.6
pydantic==2.9.2
pydantic-settings==2.6.1
python-multipart==0.0.9
requests==2.32.3
pandas==2.2.3
openpyxl==3.1.5
watchdog==5.0.3
pyyaml==6.0.2
